@@ -4,19 +4,20 @@ class MyButton extends StatelessWidget {
   final VoidCallback? onPressed;
   final String name;
 
-  const MyButton({required this.name, this.onPressed, Key? key}) : super(key: key);
+  const MyButton({required this.name, this.onPressed, Key? key})
+      : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return SizedBox(
       height: 45,
       width: double.infinity,
-      child: RaisedButton(
+      child: ElevatedButton(
         child: Text(
           name,
           style: const TextStyle(color: Colors.white),
         ),
-        color: const Color(0xff746bc9),
+        // color: const Color(0xff746bc9),
         onPressed: onPressed,
       ),
     );
