@@ -1,4 +1,4 @@
-package com.example.cau_vong_store
+package com.example.cauvongstore_mobile
 
 import io.flutter.embedding.android.FlutterActivity
 
