@@ -1,11 +1,10 @@
-import 'package:cau_vong_store/src/screen/login/login_page.dart';
+import 'package:cauvongstore_mobile/src/screen/login/login_page.dart';
 import 'package:flutter/material.dart';
 
 // import 'base/di/locator.dart';
 import 'resources/app_color.dart';
 
 class MyApp extends StatelessWidget {
-
   static GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
 
   static BuildContext get context => navigatorKey.currentContext!;
