@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:cau_vong_store/src/resources/app_color.dart';
+import 'package:cauvongstore_mobile/src/resources/app_color.dart';
 
 class RoundedButton extends StatelessWidget {
   final String text;
